client/requirements.txt
blinker==1.9.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
Flask==3.0.0
Flask-Cors==4.0.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
MouseInfo==0.1.3
Pillow==10.1.0
psutil==5.9.6
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
PyMsgBox==2.0.1
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==1.0.1
python-dotenv==1.0.0
python3-xlib==0.15
pytweening==1.2.0
requests==2.31.0
urllib3==2.5.0
Werkzeug==3.1.3
